+++ src/Models/CourseBlockProgress.php
@@ -17,22 +17,38 @@ class CourseBlockProgress extends Pivot
 
     protected $appends = [
         'time_left',
+        'progress_percentage'
     ];
 
     public function getTimeLeftAttribute()
     {
-        $videoLength = $this->courseBlock->getVideos()->pluck('video_duration')->sum();
+        $videoLength = Video::where('model_type', config('summit.course_block_model'))->where('model_id', $this->course_block_id)->sum('video_duration');
+
+        if ($videoLength <= 0) {
+            return '0 seconds';
+        }
 
         return CarbonInterval::seconds($videoLength - $this->progress)->cascade()->forHumans();
     }
 
+    public function getProgressPercentageAttribute()
+    {
+        $videoLength = Video::where('model_type', config('summit.course_block_model'))->where('model_id', $this->course_block_id)->sum('video_duration');
+
+        if ($videoLength <= 0) {
+            return '0%';
+        }
+
+        return round($this->progress / $videoLength) . '%';
+    }
+
     public function courseBlock()
     {
-        return $this->belongsTo(CourseBlock::class);
+        return $this->belongsTo(config('summit.course_block_model'), 'course_block_id');
     }
 
     public function user()
     {
-        return $this->belongsTo(User::class);
+        return $this->belongsTo(User::class, 'user_id');
     }
 }